+++ src/main/java/com/brp/service/impl/CompanyServiceImpl.java
@@ -44,7 +44,7 @@ public class CompanyServiceImpl implements CompanyService{
 
 	@Override
 	public CompanyEntity getCompanyById(Long id) {
-		CompanyEntity company = companyMapper.getCompanyById(id);
+		CompanyEntity company = companyMapper.getCompanyById(id.intValue());
 		
 		return company;
 	}
